+++ WizBot.Core/Modules/Music/Common/Impl/MusicPlayer.cs
@@ -1,6 +1,7 @@
 ﻿#nullable enable
 using System;
 using System.Collections.Generic;
+using System.ComponentModel;
 using System.Diagnostics;
 using System.Linq;
 using System.Runtime.CompilerServices;
@@ -9,6 +10,7 @@ using System.Threading;
 using System.Threading.Tasks;
 using Ayu.Discord.Voice;
 using WizBot.Common;
+using WizBot.Core.Services.Database.Models;
 using WizBot.Extensions;
 using WizBot.Modules.Music;
 using Serilog;
@@ -19,13 +21,11 @@ namespace WizBot.Core.Modules.Music
     {
         private readonly VoiceClient _vc = new VoiceClient(frameDelay: FrameDelay.Delay20);
 
-        public bool IsKilled { get; set; }
-        public bool IsStopped { get; set; }
-        public bool IsPaused { get; set; }
-        public bool IsRepeatingCurrentSong { get; set; }
-        public bool IsAutoDelete { get; set; }
-        public bool IsRepeatingQueue { get; set; } = true;
-
+        public bool IsKilled { get; private set; }
+        public bool IsStopped { get; private set; }
+        public bool IsPaused { get; private set; }
+        public PlayerRepeatType Repeat { get; private set; }
+        
         public int CurrentIndex => _queue.Index;
         
         public float Volume => _volume;
@@ -87,7 +87,7 @@ namespace WizBot.Core.Modules.Music
                     _ = _proxy.StartSpeakingAsync();
 
                     _ = OnStarted?.Invoke(this, track, index);
-                    
+
                     // make sure song buffer is ready to be (re)used
                     _songBuffer.Reset();
 
@@ -131,7 +131,7 @@ namespace WizBot.Core.Modules.Music
                             // if song is finished
                             if (result is null)
                                 break;
-                            
+
                             if (result is true)
                             {
                                 // wait for slightly less than the latency
@@ -160,7 +160,7 @@ namespace WizBot.Core.Modules.Music
                                     await Task.Delay(200, cancellationToken);
                                     continue;
                                 }
-                                
+
                                 Log.Warning("Can't send data to voice channel");
 
                                 IsStopped = true;
@@ -175,6 +175,12 @@ namespace WizBot.Core.Modules.Music
                         }
                     }
                 }
+                catch (Win32Exception)
+                {
+                    IsStopped = true;
+                    Log.Error("Please install ffmpeg and make sure it's added to your " +
+                              "PATH environment variable before trying again");
+                }
                 catch (OperationCanceledException)
                 {
                     Log.Information("Song skipped");
@@ -187,12 +193,14 @@ namespace WizBot.Core.Modules.Music
                 {
                     // turn off green in vc
                     trackCancellationSource.Cancel();
-                    HandleQueuePostTrack();
-                    
-                    _ = _proxy.StopSpeakingAsync();
+
                     _ = OnCompleted?.Invoke(this, track);
                     
+                    HandleQueuePostTrack();
                     _skipped = false;
+                    
+                    _ = _proxy.StopSpeakingAsync();;
+                    
                     await Task.Delay(100);
                 }
             }
@@ -221,33 +229,28 @@ namespace WizBot.Core.Modules.Music
                 return;
             }
 
-            if (IsRepeatingCurrentSong || IsStopped)
-                return;
+            var (repeat, isStopped) = (Repeat, IsStopped);
 
-            // autodelete is basically advance, so if it's enabled
-            // don't advance
-            if(IsAutoDelete)
-                _queue.RemoveCurrent();
+            if (repeat == PlayerRepeatType.Track || isStopped)
+                return;
             
             // if queue is being repeated, advance no matter what
-            if (!IsRepeatingQueue)
+            if (repeat == PlayerRepeatType.None)
             {
                 // if this is the last song,
                 // stop the queue
                 if (_queue.IsLast())
                 {
                     IsStopped = true;
+                    OnQueueStopped?.Invoke(this);
                     return;
                 }
                 
-                if(!IsAutoDelete)
-                    _queue.Advance();
-                
+                _queue.Advance();
                 return;
             }
             
-            if(!IsAutoDelete)
-                _queue.Advance();
+            _queue.Advance();
         }
 
         
@@ -326,6 +329,11 @@ namespace WizBot.Core.Modules.Music
             _queue.EnqueueMany(tracks, queuer);
         }
 
+        public void SetRepeat(PlayerRepeatType type)
+        {
+            Repeat = type;
+        }
+
         public void ShuffleQueue()
         {
             _queue.Shuffle(_rng);
@@ -396,10 +404,7 @@ namespace WizBot.Core.Modules.Music
 
             return true;
         }
-
-        public bool ToggleRcs() => IsRepeatingCurrentSong = !IsRepeatingCurrentSong;
-        public bool ToggleRpl() => IsRepeatingQueue = !IsRepeatingQueue;
-        public bool ToggleAd() => IsAutoDelete = !IsAutoDelete;
+        
         public bool TogglePause() => IsPaused = !IsPaused;
         public IQueuedTrackInfo? MoveTrack(int from, int to) => _queue.MoveTrack(from, to);
 
@@ -408,6 +413,7 @@ namespace WizBot.Core.Modules.Music
             IsKilled = true;
             OnCompleted = null;
             OnStarted = null;
+            OnQueueStopped = null;
             _queue.Clear();
             _songBuffer.Dispose();
             _vc.Dispose();
@@ -415,5 +421,6 @@ namespace WizBot.Core.Modules.Music
 
         public event Func<IMusicPlayer, IQueuedTrackInfo, Task>? OnCompleted;
         public event Func<IMusicPlayer, IQueuedTrackInfo, int, Task>? OnStarted;
+        public event Func<IMusicPlayer, Task>? OnQueueStopped;
     }
 }