+++ WizBot.Core/Modules/Music/Services/IMusicService.cs
@@ -1,8 +1,10 @@
-﻿using System.Collections.Generic;
+﻿#nullable enable
+using System.Collections.Generic;
 using System.Threading.Tasks;
 using Discord;
 using WizBot.Core.Common;
 using WizBot.Core.Modules.Music;
+using WizBot.Core.Services.Database.Models;
 
 namespace WizBot.Modules.Music.Services
 {
@@ -21,16 +23,17 @@ namespace WizBot.Modules.Music.Services
         /// <param name="voiceChannelId">Id of the voice channel</param>
         public Task JoinVoiceChannelAsync(ulong guildId, ulong voiceChannelId);
         
-        IMusicPlayer GetOrCreateMusicPlayer(ITextChannel contextChannel);
+        Task<IMusicPlayer?> GetOrCreateMusicPlayerAsync(ITextChannel contextChannel);
         bool TryGetMusicPlayer(ulong guildId, out IMusicPlayer musicPlayer);
-        void SetDefaultVolume(ulong guildId, int val);
         Task<int> EnqueueYoutubePlaylistAsync(IMusicPlayer mp, string playlistId, string queuer);
         Task EnqueueDirectoryAsync(IMusicPlayer mp, string dirPath, string queuer);
         Task<int> EnqueueSoundcloudPlaylistAsync(IMusicPlayer mp, string playlist, string queuer);
-        Task<IUserMessage> SendToOutputAsync(ulong guildId, EmbedBuilder embed);
-        bool SetMusicChannel(ulong guildId, ulong channelId);
-        void UnsetMusicChannel(ulong guildId);
+        Task<IUserMessage?> SendToOutputAsync(ulong guildId, EmbedBuilder embed);
         Task<bool> PlayAsync(ulong guildId, ulong voiceChannelId);
         Task<IList<(string Title, string Url)>> SearchVideosAsync(string query);
+        Task<bool> SetMusicChannelAsync(ulong guildId, ulong? channelId);
+        Task SetRepeatAsync(ulong guildId, PlayerRepeatType repeatType);
+        Task SetVolumeAsync(ulong guildId, int value);
+        Task<bool> ToggleAutoDisconnectAsync(ulong guildId);
     }
 }