+++ src/ayu/Ayu.Discord.Voice/SongBuffer.cs
@@ -46,6 +46,7 @@ Please install and configure FFMPEG to play music.
 Check the guides for your platform on how to setup ffmpeg correctly:
     Windows Guide: https://goo.gl/OjKk8F
     Linux Guide:  https://goo.gl/ShjCUo");
+                throw;
             }
             catch (OperationCanceledException)
             {